requirements.txt
# Astro Runtime includes the following pre-installed providers packages
apache-airflow-providers-postgres==5.4.0

# https://pypi.org/project/astronomer-providers/
astronomer-providers==1.15.2

# https://pypi.org/project/minio/
minio==7.1.14
